refactor: Allow decimal values for beverage amounts and refactor dialog logic
This commit refactors the beverage tracking feature to support decimal values for amounts, improving precision. It also decouples the "Add Beverage" dialog logic from `BeverageOverviewViewModel` into a new, dedicated `AddBeverageDialogViewModel`.

### Key Changes:

-   **Decimal Support for Beverages:**
    -   Updated `Beverage`, `BeverageEntity`, and related classes to use `Double` for the `amount` field instead of `UInt` or `Int`.
    -   Changed the input field for beverage amounts in `ProfileView.kt` and `AddBeverageDialog.kt` to use `KeyboardType.Decimal`.
    -   Modified `FluidUnit` and `FluidUnitConverter` to perform conversions with `Double`.
    -   Updated string resources (`strings.xml`) for English and German to format beverage quantities as decimals (`%1$.2f`).
    -   Adjusted seed data in `FitBeDatabase.kt` to use `Double` for beverage amounts.

-   **Refactoring "Add Beverage" Dialog:**
    -   Created a new `AddBeverageDialogViewModel.kt` to manage the state and validation logic for the "Add Beverage" dialog.
    -   Migrated dialog state management (amount, name, unit, date, errors) from `BeverageOverviewViewModel` to the new `AddBeverageDialogViewModel`.
    -   Introduced `BeverageDialogUiState` to hold the state for the dialog.
    -   The `BeverageOverviewView` now uses a local state (`showAddBeverageDialog`) to control the dialog's visibility and is injected with both `BeverageOverviewViewModel` and the new `AddBeverageDialogViewModel`.
    -   Updated DI in `Modules.kt` to provide `AddBeverageDialogViewModel`.

-   **Improved Validation:**
    -   Replaced `PositiveNumberValidator` with `PositiveDecimalValidator` for beverage amount validation.
    -   The `BeverageValidator` now checks `Double` values.
    -   `AddBeverageDialogViewModel` now performs validation against the total daily beverage intake to prevent logging excessive amounts (over 5000ml). This check considers the currently entered amount and previously logged beverages for the selected day.

-   **Bug Fix in Profile Name Validation:**
    -   Fixed a bug in `ProfileViewModel.onNameChanged` where the UI state was not immediately updated upon input, causing a laggy user experience. The `inputName` is now updated instantly before the asynchronous validation runs.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/components/validators/BeverageValidator.kt
@@ -1,7 +1,7 @@
 package org.darthacheron.fitbe.components.validators
 
 class BeverageValidator {
-    fun validate(value: UInt?): Boolean {
-        return value == null || value <= 5000u
+    fun validate(value: Double?): Boolean {
+        return value == null || value <= 5000
     }
 }
